Add files via upload
# Project Guide: Integrating Static Analysis Tools with Defender for DevOps

This document serves as a comprehensive guide on integrating static analysis tools like Bandit, ESLint, and Credscan into the development lifecycle with Microsoft Defender for DevOps. It details the steps taken to set up security checks in Azure DevOps and GitHub Actions, aimed at identifying and mitigating potential security vulnerabilities early in the development process. Additionally, this guide explains how to utilize the DevOps Security Workbook for ongoing security monitoring and reporting.

You are welcome to use this guide for ideas on how to implement similar security measures in your own DevOps pipeline.